+++ services/gastown/src/dos/Town.do.ts
@@ -2747,7 +2747,6 @@ export class TownDO extends DurableObject<Env> {
 
 
 
-
   /**
    * Create an open bead with the given labels, without arming the reconciler alarm.
    * The caller is responsible for including `gt:held` in the labels if the bead
